Use a different kind of parser (Push) for reading GraphML files. This might solve a mysterious bug reported by Kristian Ovaska. If called from R, the parser sometimes failed for some files. Very hard to reproduce.
However it wasn't clear for me how the actual read worked with the
I/O parser, there was actually two ways to read from the file, first 
with the read callback and then directly via fread and pass the 
data to xmlParseChunk(). The current method seems to work and 
at least i understand what is happening here.

Patches applied:

 * csardi@szaffi2/igraph--main--0.5--patch-1
   Use a different kind of parser (Push) for reading GraphML files.