Isomorphism and subgraph isomorphism updates There are two functions now, one for isomorphism, another one to subgraph isomorphism. These functions call a handler with every mapping found, then continue backtracking only if the handler returns TRUE and stop otherwise.
The other functions for the 1) decision, 2) counting and 3) calculation
of the isomorphisms use these handlers.

Actually the two functions for isomorphisms and subgraph isomorphisms
could be merged easily, only two of three minor details differ, but
this way it is possible to build in different heuristics for example.
They still might be merged later with keeping the current API.

Patches applied:

 * csardi@szaffi2/igraph--main--0.5--patch-46
   Isomorphism and subgraph isomorphism updates