R: fix attribute handling to (partially) solve the UNPROTECT issue.
The problem is that attributes must be temporarily 
PROTECT-ed, while in the igraph_t. If an error happens,
then the freshly PROTECT-ed stuff in UNPROTECT-ed 
automatically by the R error handling machinery. 

Plus, the igraph error handler destroys the unneeded 
leftover stuff, which is fine, but it also calls 
igraph_destroy of the graph(s), and this wants to destroy 
the attributes and calls UNPROTECT again. This is a problem.

So now we have a flag in the attributes that shows whether it is 
safe to call UNPROTECT manually or not.

This is not a complete solution, however. Plus, other functions 
are wrong as well, e.g. delete_edges, etc. This is not a serious
problem, since errors cannot really happen for these functions
right now, or if they do, that's already too bad anyway (e.g.
we are out of memory).

For the complete solution part of the R attribute handler
should be rewritten.