DYN: index edge list according to birth, too
More precisely oi is indexed according to
1. source vertex
2. birth time of edge
3. target vertex
and ii is index according to
1. target vertex
2. birth time of edge
3. soruce vertex.

This allows finding the neighbors and incident edges of a
vertex more efficiently. One needs to be careful, though,
because neighbors() and incident() must still return the vertices
and edges in increasing vertex id order. Several algorithms
rely on this, and they should work on a sliced (=now set to finite
time point) graph.