global reaching centrality function
I've added a function called GRC. 

GRC is the global reaching centrality of the graph, which is a measure of hierarchy.1 indicates a high degree of centrality, while 0 indicates a low degree of centrality. The graph must be a directed for the measure to be valid. 

It is as described in Mones, Enys, Lilla Vicsek, and Tamás Vicsek. "Hierarchy measure for complex networks." PloS one 7, no. 3 (2012): e33799.