Fix #1311 by checking if number of edges is zero
In igraph_community_edge_betweenness, if the number of edges in the
graph is zero, and an initialized pointer is passed for the edge
betweenness variable, then the array will be indexed at -1, which
causes memory corruption in the chunk header, making the pointer
invalid. This will segfault when free is called on the pointer.