Add layout_fruchterman_reingold unit tests
Part of #1592

This one is hard to test. I mainly put in some values to see of they
looked plausible. It also just seems very well guarded against memory
problems, which is great :)
It is easy to get into obviously suboptimal results by for example
turning the temperature up to ridiculously high values, but that seems
reasonable.

If there are suggestions for different kinds of tests I will of course
gladly add them.