Add igraph_shortest_paths_johnson unit tests
Part of #1592

I changed the output of NaN's and Infs when they're printed with %8g,
for better alignment. This also required fixing modularity_matrix
output.

Bugfix: If there are no edges, there are no negative weights. This
check was not correct before.

Bugfix: When selecting a subset of to and from vertices, the `to`
indices were incorrect when reweighting the results. It used the indices
of the matrix for the whole graph instead of the vertex subset.