Add solve_lsap unit tests
Part of #1592

I only noticed this was in /src/internal after adding tests.
It is not used internally. I therefore only did the minimal amount
of work to make this work correctly in some cases, and dissalow other
cases. If we want this to be part of public igraph, I can fix some
more things, like missing IGRAPH_CHECK calls, the malloc and calloc
calls, and the superfluous n parameter which can be taken from the matrix.