Draft PR: skeleton for rich-club coefficient
This is a draft PR with the basic skeleton for the rich club calculations in properties/rich_club.c (function declarations have been added to igraph_structural.h, and the new .c file to CMakeLists.txt). The parameters and details are still being worked out, especially the options for weighted graphs, etc. - would love any and all feedback on this!